% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instant.R
\name{ct_attr}
\alias{ct_attr}
\title{Make a function that gets variable-names}
\usage{
ct_attr(attr = c("ch_id", "ph"))
}
\arguments{
\item{attr}{character indicating which attribute of channels to use as names}
}
\value{
function to get variable-names from config
}
\description{
This is essentially a wrapper for \code{\link[=ct_channel_name]{ct_channel_name()}},
allowing you to delay the evaluation of the comtrade object.
}
\examples{
  ch_name <- ct_attr("ch_id")
  ch_name(keating_1999)
}
