% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_config.R
\name{ct_config_analog_channel}
\alias{ct_config_analog_channel}
\title{Create data-frame for analog channels}
\usage{
ct_config_analog_channel(.data, ch_id = NULL, ph = NULL, ccbm = NULL, uu,
  a, b, skew = NULL, min = -99999L, max = 99999L, primary = 1,
  secondary = 1, PS = "P")
}
\arguments{
\item{.data}{data frame}

\item{PS}{}
}
\value{
\code{tbl_df}
}
\description{
Create data-frame for analog channels
}
