% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tall.R
\name{ct_data_tall}
\alias{ct_data_tall}
\alias{ct_data_tall.default}
\alias{ct_data_tall.comtrade}
\title{Make a Tall Data Frame}
\usage{
ct_data_tall(ct, ...)

\method{ct_data_tall}{default}(ct, ...)

\method{ct_data_tall}{comtrade}(ct, ...)
}
\arguments{
\item{ct}{comtrade object (see \code{\link[=comtrade]{comtrade()}})}
}
\value{
data frame
}
\description{
The COMTRADE format for the data is a wide data frame, it could be handy
to have a tall format.
}
\details{
Here are the variables:

\describe{
\item{\code{type}}{character, type of channel: analog or digital}
\item{\code{channel_number}}{integer, number of channel within \code{type}}
\item{\code{n}}{integer, number of observation within \code{type} and \code{channel_number}}
\item{\code{timestamp}}{integer, non-scaled timestamp}
\item{\code{value}}{integer, non-scaled value}
}
}
\keyword{internal}
