% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{ct_example}
\alias{ct_example}
\title{path to example}
\usage{
ct_example(path)
}
\arguments{
\item{path}{character, name of file}
}
\value{
path to the file
}
\description{
Currently there are three example-files that all come from the same waveform.
The two files, \code{"keating_1999.CFG"} and \code{"keating_1999.DAT"} are the contents of
\code{"keating_1999.zip"}.
}
\examples{
\dontrun{
  ct_example("keating_1999.CFG")
  ct_example("keating_1999.DAT")
  ct_example("keating_1999.zip")
}
  ct_example("keating_1999.zip") \%>\%
  ct_read_zip() \%>\%
  ct_instant()
}
