% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_config.R
\name{ct_read_config}
\alias{ct_read_config}
\title{read config file}
\usage{
ct_read_config(file)
}
\arguments{
\item{file}{character, single line, path to a file or a single string}
}
\value{
list, specified in \code{vignette("config")}
}
\description{
read config file
}
