% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{ct_read_data}
\alias{ct_read_data}
\title{Read data file}
\usage{
ct_read_data(file, n_analog, n_digital, n_max = Inf)
}
\arguments{
\item{file}{character, single line, path to a file or a single string}

\item{n_analog}{integer, number of analog channels}

\item{n_digital}{integer, number of digital channels}

\item{n_max}{integer, maximum number of records to read}
}
\value{
tibble (data frame)
}
\description{
In practice, it may be easier to use \code{\link[=ct_read_data_config]{ct_read_data_config()}}.
}
\examples{
  config <- ct_example("keating_1999.CFG") \%>\% ct_read_config()
  ct_example("keating_1999.DAT") \%>\%
    ct_read_data(
      n_analog = config[["##A"]],
      n_digital = config[["##D"]],
      n_max = max(config$sampling_rate$endsamp)
    )
}
\seealso{
\code{\link[=ct_read_data_config]{ct_read_data_config()}}
}
