% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_zip.R
\name{ct_read_zip}
\alias{ct_read_zip}
\title{Read a zip file to make a comtrade object}
\usage{
ct_read_zip(file)
}
\arguments{
\item{file}{character, path to a zip file}
}
\value{
comtrade object
}
\description{
Read a zip file to make a comtrade object
}
\examples{
  ct_example("keating_1999.zip") \%>\%
  ct_read_zip()
}
