% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_per_cycle.R
\name{ct_sample_per_cycle_single}
\alias{ct_sample_per_cycle_single}
\title{Samples per cycle.}
\usage{
ct_sample_per_cycle_single(ct)
}
\arguments{
\item{ct}{\code{comtrade} S3 object constructed using \code{\link[=comtrade]{comtrade()}}}
}
\value{
\code{integer} estimate of number of samples per cycle
}
\description{
This is a convenience function for those cases where the comtrade file has
\strong{exactly one} sampling frequency.
}
\details{
This function uses some simple heuristics to estimate the sampling frequency,
assuming that it will be close to the line-frequency multiplied by some
positive-integer power of two.
}
