% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{ct_serialize_config}
\alias{ct_serialize_config}
\title{Serialize a config object}
\usage{
ct_serialize_config(config)
}
\arguments{
\item{config}{\code{list}, comtrade configuration}
}
\value{
\code{character}
}
\description{
We are confined to the 1999 revision of comtrade
}
