% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{ctread_config}
\alias{ctread_config}
\title{read config file}
\usage{
ctread_config(file)
}
\arguments{
\item{file}{character, single line, path to a file or a single string}
}
\value{
list, specified in \code{vignette("header")}
}
\description{
read config file
}
