% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instant.R
\name{fn_scale_channel}
\alias{fn_scale_channel}
\title{Get function to scale analog-channel}
\usage{
fn_scale_channel(config, i_analog)
}
\arguments{
\item{config}{list, configuration generated using \code{\link[=ct_read_config]{ct_read_config()}}}

\item{i_analog}{integer, analog channel-number}
}
\value{
function that takes a value and returns a scaled value
}
\description{
Given a config file and an analog-channel number,
return a function to scale the value.
}
\keyword{internal}
