% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instant.R
\name{fn_scale_side}
\alias{fn_scale_side}
\title{Get function to scale side}
\usage{
fn_scale_side(config, side = c("primary", "secondary"))
}
\arguments{
\item{config}{list, configuration generated using \code{\link[=ct_read_config]{ct_read_config()}}}

\item{side}{character to indicate if we want primary or
secondary values}
}
\description{
Given a config file and a side, return a function to scale a value
}
\keyword{internal}
