% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{keating_1999}
\alias{keating_1999}
\title{Keating waveform, 1999 format}
\format{a comtrade object, a list with two members,
generated using \code{\link[=comtrade]{comtrade()}}
\describe{
\item{config}{comtrade config object}
\item{data}{cpmtrade data object}
}}
\usage{
keating_1999
}
\description{
This is a waveform capture from Schneider Electric's Victoria-Keating facility.
}
\keyword{datasets}
