% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_config.R
\name{parse_config}
\alias{parse_config}
\title{parse into a data frame}
\usage{
parse_config(text, ...)
}
\arguments{
\item{text}{character, vector of strings}

\item{...}{other args passed to \code{\link[readr:read_csv]{readr::read_csv()}}}
}
\value{
tibble
}
\description{
parse into a data frame
}
\keyword{internal}
