% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comtrade.R
\name{comtrade}
\alias{comtrade}
\title{comtrade constructor}
\usage{
comtrade(config, data, header = NULL, info = NULL)
}
\arguments{
\item{config}{list, configuration generated using \code{\link[=ct_read_config]{ct_read_config()}}}

\item{data}{data frame, data generated using \code{\link[=ct_read_data]{ct_read_data()}}}

\item{header}{character, header (not yet allocated)}

\item{info}{character, information (not yet alllocated)}
}
\value{
comtrade object
}
\description{
comtrade constructor
}
