% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instant.R
\name{ct_channel_name}
\alias{ct_channel_name}
\title{Helper function for channel names}
\usage{
ct_channel_name(ct, attr = c("ch_id", "ph"))
}
\arguments{
\item{ct}{comtrade object (see \code{\link[=comtrade]{comtrade()}})}

\item{attr}{character indicating which attribute of channels to use as names}
}
\value{
list with two members: \code{analog} and \code{digital} -
each of which is a vector of character column-names
}
\description{
Helper function for channel names
}
\examples{
  ct_channel_name(keating_1999, attr = "ch_id")
  ct_channel_name(keating_1999, attr = "ph")
}
