% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{ct_read_data}
\alias{ct_read_data}
\title{read data file}
\usage{
ct_read_data(file, n_analog, n_digital, n_max = Inf)
}
\arguments{
\item{file}{character, single line, path to a file or a single string}

\item{n_analog}{integer, number of analog channels}

\item{n_digital}{integer, number of digital channels}

\item{n_max}{integer, maximum number of records to read}
}
\value{
tibble, specified in \code{vignette("data")}
}
\description{
read data file
}
