% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{ct_read_data_config}
\alias{ct_read_data_config}
\title{read data file}
\usage{
ct_read_data_config(file, config)
}
\arguments{
\item{file}{character, single line, path to a file or a single string}

\item{config}{list created using \code{\link[=ct_read_config]{ct_read_config()}}}
}
\value{
tibble, specified in \code{vignette("data")}
}
\description{
read data file
}
