% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor.R
\name{ct_sample_per_cycle}
\alias{ct_sample_per_cycle}
\alias{ct_sampling_frequency_nominal}
\alias{ct_electrical_frequency_nominal}
\alias{ct_sampling_frequency}
\title{Accessor functions}
\usage{
ct_sample_per_cycle(ct)

ct_sampling_frequency_nominal(ct)

ct_electrical_frequency_nominal(ct)

ct_sampling_frequency(ct)
}
\arguments{
\item{ct}{\code{comtrade} S3 object constructed using \code{\link[=comtrade]{comtrade()}}}
}
\value{
An \code{integer}, estimate of number of samples per cycle.

A \code{numeric}, estimate of nominal sampling-rate (Hz.).

A \code{numeric}, electrical frequency (Hz.)

A \code{numeric}, electrical frequency (Hz.)
}
\description{
This is a convenience function for those cases where the comtrade file has
\strong{exactly one} sampling frequency.
}
\details{
This function uses some simple heuristics to estimate the sampling frequency,
assuming that it will be close to the line-frequency multiplied by some
positive-integer power of two.
}
\examples{
  library("ieeecomtrade")

  ct_sample_per_cycle(keating_1999)
  ct_sampling_frequency_nominal(keating_1999)
}
